Switched lib XML parsing to accept huge files as this can happen with base64-encoded raw images.
The parser of course trips over this even if the image never gets exported.

I had first thought this was an Inkscape bug, but it's a libxml feature to
protect from denial of service attakcs from huge input, which should never
be an issue with this code, as it is assumed to only get trusted input
anyway.