fix: checks for multiple comics ignored cached comics
When a comic has been cached, it's not checked again. This is because it has
been checked before, and we don't cache all its meta information (e.g.,
transcript), so some checks would have to go back to the original svg,
negating the caching effects.

This led to the situation that cached comics were never passed to the
Checks, and any Check that needed to look at multiple comics (e.g., to
detect a typo in series) only saw the new comic, triggering a series warning
even when there were multiple comics in the series. Re-running removed the
waning as now all comics were cached and hence none checked.

This commit also adds a Check base class to DRY up the code and flesh aout
the Check interface.