refact: moving HTML link generation to its own module
This also changes the output name to make it clearer that this is HTML
centric (docbook or latex will need their own linking modules).

Adds the references to the linking comic, to follow the general pattern of
treating Comic as a data collection object that the templates can work off.