test: configuring Test::Pod::No404s to ignore a particular link
I've submitted a pull request for a ignore file for Test::Pod::No404s.

The given URL needs to be ignored cause it returns a405 on curl et al, but
triggers a redirect in the browser to log in to your Twitter account. Of
course Test::Pod::No404s sees the 405 as an error and fails the test, even
when in real life there is no broken link.

To add insult to injury, it has a fairly low rate limit.