fix(Comic): removed special treatment for Thursday
The old code assumed comics were published on Fridays, and when running on
Thursday, would already include next Friday's comic as if it was published.
This was a performance optimization hack, so that a cron job could run at
23:00 on Thursday and convert everything, and then another cron job at
midnight on Friday could just upload everything. This is obsolete for a
while now as only changed comics are converted anyway and the whole process
is fairly quick.

Removing this special treatment also removes the hard-coded assumption that
comics are published on Fridays.